✨ Feat: 와이리 가입하기 배너 #46
