+++ .github/workflows/ci.yml
@@ -17,7 +17,6 @@ jobs:
         uses: actions/setup-node@v3
         with:
           node-version: '20'
-          cache: npm # npm 캐시 사용
 
       - name: Install dependencies
         run: npm install
